Training/SLMorg/requirements.txt
wandb
tqdm
transformers
tokenizers
pandas
pyarrow
torchinfo
nvitop